change type of state
